Decouple ContentLocalization from Autoroute.
