+++ src/OrchardCore.Modules/OrchardCore.Autoroute/Services/AutorouteAliasProvider.cs
@@ -1,5 +1,6 @@
 using System.Threading.Tasks;
 using OrchardCore.ContentManagement;
+using OrchardCore.ContentManagement.Routing;
 
 namespace OrchardCore.Autoroute.Services
 {
