+++ src/OrchardCore.Modules/OrchardCore.ContentLocalization/OrchardCore.ContentLocalization.csproj
@@ -19,7 +19,6 @@
     <ProjectReference Include="..\..\OrchardCore\OrchardCore.Module.Targets\OrchardCore.Module.Targets.csproj" />
     <ProjectReference Include="..\..\OrchardCore\OrchardCore.ResourceManagement\OrchardCore.ResourceManagement.csproj" />
 
-    <ProjectReference Include="..\OrchardCore.Autoroute\OrchardCore.Autoroute.csproj" />
     <ProjectReference Include="..\OrchardCore.Localization\OrchardCore.Localization.csproj" />
   </ItemGroup>
