+++ src/OrchardCore.Modules/OrchardCore.ContentLocalization/Records/Migrations.cs
@@ -7,7 +7,7 @@ namespace OrchardCore.ContentLocalization.Records
 {
     public class Migrations : DataMigration
     {
-        IContentDefinitionManager _contentDefinitionManager;
+        private readonly IContentDefinitionManager _contentDefinitionManager;
 
         public Migrations(IContentDefinitionManager contentDefinitionManager)
         {
